requirements.txt
Django==3.2.16
djangorestframework==3.14.0
django-environ==0.7.0
psycopg2-binary==2.9.1
segyio==1.9.9
importlib-metadata==4.13.0
celery==4.4.7
redis
redis-server
numpy==1.21.6
pandas==1.3.5
django-cors-headers==3.13.0
django-redis
django-crontab